# Optimize Empty Chunk Equality