--- a/core/shared/src/main/scala/zio/Chunk.scala
+++ b/core/shared/src/main/scala/zio/Chunk.scala
@@ -46,53 +46,47 @@ sealed abstract class Chunk[+A] extends ChunkLike[A] with Serializable { self =>
    * Returns the concatenation of this chunk with the specified chunk.
    */
   final def ++[A1 >: A](that: Chunk[A1]): Chunk[A1] =
-    if (isEmpty) {
-      that
-    } else if (that.isEmpty) {
-      this
-    } else {
-      (self, that) match {
-        case (Chunk.AppendN(start, buffer, bufferUsed, _), that) =>
-          val chunk = Chunk.fromArray(buffer.asInstanceOf[Array[A1]]).take(bufferUsed)
-          start ++ chunk ++ that
-        case (self, Chunk.PrependN(end, buffer, bufferUsed, _)) =>
-          val chunk = Chunk.fromArray(buffer.asInstanceOf[Array[A1]]).takeRight(bufferUsed)
-          self ++ chunk ++ end
-        case (self, Chunk.Empty) => self
-        case (Chunk.Empty, that) => that
-        case (self, that) =>
-          val diff = that.depth - self.depth
-          if (math.abs(diff) <= 1) Chunk.Concat(self, that)
-          else if (diff < -1) {
-            if (self.left.depth >= self.right.depth) {
-              val nr = self.right ++ that
+    (self, that) match {
+      case (Chunk.AppendN(start, buffer, bufferUsed, _), that) =>
+        val chunk = Chunk.fromArray(buffer.asInstanceOf[Array[A1]]).take(bufferUsed)
+        start ++ chunk ++ that
+      case (self, Chunk.PrependN(end, buffer, bufferUsed, _)) =>
+        val chunk = Chunk.fromArray(buffer.asInstanceOf[Array[A1]]).takeRight(bufferUsed)
+        self ++ chunk ++ end
+      case (self, Chunk.Empty) => self
+      case (Chunk.Empty, that) => that
+      case (self, that) =>
+        val diff = that.depth - self.depth
+        if (math.abs(diff) <= 1) Chunk.Concat(self, that)
+        else if (diff < -1) {
+          if (self.left.depth >= self.right.depth) {
+            val nr = self.right ++ that
+            Chunk.Concat(self.left, nr)
+          } else {
+            val nrr = self.right.right ++ that
+            if (nrr.depth == self.depth - 3) {
+              val nr = Chunk.Concat(self.right.left, nrr)
               Chunk.Concat(self.left, nr)
             } else {
-              val nrr = self.right.right ++ that
-              if (nrr.depth == self.depth - 3) {
-                val nr = Chunk.Concat(self.right.left, nrr)
-                Chunk.Concat(self.left, nr)
-              } else {
-                val nl = Chunk.Concat(self.left, self.right.left)
-                Chunk.Concat(nl, nrr)
-              }
-            }
-          } else {
-            if (that.right.depth >= that.left.depth) {
-              val nl = self ++ that.left
-              Chunk.Concat(nl, that.right)
-            } else {
-              val nll = self ++ that.left.left
-              if (nll.depth == that.depth - 3) {
-                val nl = Chunk.Concat(nll, that.left.right)
-                Chunk.Concat(nl, that.right)
-              } else {
-                val nr = Chunk.Concat(that.left.right, that.right)
-                Chunk.Concat(nll, nr)
-              }
+              val nl = Chunk.Concat(self.left, self.right.left)
+              Chunk.Concat(nl, nrr)
             }
           }
-      }
+        } else {
+          if (that.right.depth >= that.left.depth) {
+            val nl = self ++ that.left
+            Chunk.Concat(nl, that.right)
+          } else {
+            val nll = self ++ that.left.left
+            if (nll.depth == that.depth - 3) {
+              val nl = Chunk.Concat(nll, that.left.right)
+              Chunk.Concat(nl, that.right)
+            } else {
+              val nr = Chunk.Concat(that.left.right, that.right)
+              Chunk.Concat(nll, nr)
+            }
+          }
+        }
     }
 
   /**
@@ -392,7 +386,7 @@ sealed abstract class Chunk[+A] extends ChunkLike[A] with Serializable { self =>
       dropping as builder.result()
     }
 
-  override final def equals(that: Any): Boolean =
+  override def equals(that: Any): Boolean =
     (self eq that.asInstanceOf[AnyRef]) || (that match {
       case that: Seq[_] => self.corresponds(that)(_ == _)
       case _            => false
@@ -2407,6 +2401,9 @@ object Chunk extends ChunkFactory with ChunkPlatformSpecific {
 
     override def apply(n: Int): Nothing =
       throw new ArrayIndexOutOfBoundsException(s"Empty chunk access to $n")
+
+    override def equals(that: Any): Boolean =
+      self eq that.asInstanceOf[AnyRef]
 
     override def foreach[B](f: Nothing => B): Unit = {
       val _ = f