We need to make checking whether a chunk is empty extremely fast instead of working around it not being fast.

Before:

```
[info] Benchmark                                      (size)   Mode  Cnt         Score        Error  Units
[info] ChunkConcatBenchmarks.balancedConcatOnce        10000  thrpt   15  90394571.169 ± 215555.543  ops/s
[info] ChunkConcatBenchmarks.balancedConcatRecursive   10000  thrpt   15      2870.790 ±     14.923  ops/s
[info] ChunkConcatBenchmarks.leftConcat0               10000  thrpt   15     56978.223 ±    134.990  ops/s
[info] ChunkConcatBenchmarks.leftConcat1               10000  thrpt   15       108.647 ±      0.726  ops/s
[info] ChunkConcatBenchmarks.leftConcat10              10000  thrpt   15      2147.022 ±      3.377  ops/s
[info] ChunkConcatBenchmarks.rightConcat0              10000  thrpt   15     56958.191 ±    161.059  ops/s
[info] ChunkConcatBenchmarks.rightConcat1              10000  thrpt   15        76.607 ±      0.487  ops/s
[info] ChunkConcatBenchmarks.rightConcat10             10000  thrpt   15      1539.452 ±      1.926  ops/s
```

After:

```
[info] ChunkConcatBenchmarks.balancedConcatOnce        10000  thrpt   15  172412535.947 ± 2139866.918  ops/s
[info] ChunkConcatBenchmarks.balancedConcatRecursive   10000  thrpt   15      14818.441 ±     491.294  ops/s
[info] ChunkConcatBenchmarks.leftConcat0               10000  thrpt   15      56922.889 ±      65.323  ops/s
[info] ChunkConcatBenchmarks.leftConcat1               10000  thrpt   15       1037.504 ±       2.270  ops/s
[info] ChunkConcatBenchmarks.leftConcat10              10000  thrpt   15      14567.367 ±      16.552  ops/s
[info] ChunkConcatBenchmarks.rightConcat0              10000  thrpt   15      56851.307 ±      95.791  ops/s
[info] ChunkConcatBenchmarks.rightConcat1              10000  thrpt   15       1025.754 ±       1.151  ops/s
[info] ChunkConcatBenchmarks.rightConcat10             10000  thrpt   15      14336.929 ±      17.053  ops/s
```